+++ laktory/metadata.py
@@ -1,0 +1,16 @@
+from typing import Union
+
+
+from laktory.models.compute.pipeline import Pipeline
+from laktory._settings import settings
+
+
+def read_metadata(pipeline: str = None) -> Union[Pipeline]:
+
+    if pipeline is not None:
+
+        filepath = f"/Workspace{settings.workspace_root}pipelines/{pipeline}.json"
+        with open(filepath, "r") as fp:
+            pl = Pipeline.model_validate_json(fp.read())
+
+        return pl