+++ laktory/models/compute/warehouse.py
@@ -4,6 +4,20 @@ from laktory.models.resources import Resources
 from laktory.models.permission import Permission
 
 
+class WarehouseCustomTag(BaseModel):
+    key: str
+    value: str
+
+
+class WarehouseTags(BaseModel):
+    custom_tags: list[WarehouseCustomTag] = []
+
+    @property
+    def pulumi_args(self):
+        import pulumi_databricks as databricks
+        return databricks.SqlEndpointTagsArgs(**self.model_dump())
+
+
 class Warehouse(BaseModel, Resources):
     cluster_size: Literal["2X-Small", "X-Small", "Small", "Medium", "Large", "X-Large", "2X-Large", "3X-Large", "4X-Large"]
     auto_stop_mins: int = None
@@ -20,7 +34,7 @@ class Warehouse(BaseModel, Resources):
     # odbc_params
     spot_instance_policy: Literal["COST_OPTIMIZED", "RELIABILITY_OPTIMIZED"] = None
     # state
-    tags: dict[str, str] = None
+    tags: WarehouseTags = None
     warehouse_type: Literal["CLASSIC", "PRO"] = None
 
     # Deployment Options