+++ laktory/resourcesengines/pulumi/cluster.py
@@ -29,85 +29,35 @@ class PulumiCluster(PulumiResourcesEngine):
             delete_before_replace=True,
         )
 
-        autoscale = None
-        if cluster.autoscale:
-            autoscale = databricks.ClusterAutoscaleArgs(
-                min_workers=cluster.autoscale.min_workers,
-                max_workers=cluster.autoscale.max_workers,
-            )
-
-        init_scripts = []
-        for i in cluster.init_scripts:
-            if i.volumes:
-                init_scripts += [
-                    databricks.ClusterInitScriptArgs(
-                        volumes=databricks.ClusterInitScriptVolumesArgs(destination=i.volumes.destination)
-                    )
-                ]
-
-            if i.workspace:
-                init_scripts += [
-                    databricks.ClusterInitScriptArgs(
-                        workspace=databricks.ClusterInitScriptWorkspaceArgs(destination=i.volumes.destination)
-                    )
-                ]
-
-        libraries = []
-        for l in cluster.libraries:
-            if l.cran:
-                libraries += [databricks.ClusterLibraryArgs(
-                    cran=databricks.ClusterLibraryCranArgs(
-                        package=l.cran.package,
-                        repo=l.cran.repo,
-                    )
-                )]
-            if l.egg:
-                libraries += [databricks.ClusterLibraryArgs(
-                    egg=l.egg
-                )]
-            if l.jar:
-                libraries += [databricks.ClusterLibraryArgs(
-                    jar=l.jar
-                )]
-            if l.maven:
-                libraries += [databricks.ClusterLibraryArgs(
-                    maven=databricks.ClusterLibraryMavenArgs(
-                        coordinates=l.maven.coordinates,
-                        exclusions=l.maven.exclusions,
-                        repo=l.maven.repo,
-                    )
-                )]
-            if l.pypi:
-                libraries += [databricks.ClusterLibraryArgs(
-                    pypi=databricks.ClusterLibraryPypiArgs(
-                        package=l.pypi.package,
-                        repo=l.pypi.repo,
-                    )
-                )]
-            if l.whl:
-                libraries += [databricks.ClusterLibraryArgs(
-                    whl=l.whl
-                )]
-
         self.cluster = databricks.Cluster(
-                f"cluster-{cluster.name}",
-                apply_policy_default_values=cluster.apply_policy_default_values,
-                autoscale=autoscale,
-                autotermination_minutes=cluster.autotermination_minutes,
-                cluster_name=cluster.name,
-                custom_tags=cluster.custom_tags,
-                data_security_mode=cluster.data_security_mode,
-                init_scripts=init_scripts,
-                libraries=libraries,
-                node_type_id=cluster.node_type_id,
-                num_workers=cluster.num_workers,
-                runtime_engine=cluster.runtime_engine,
-                single_user_name=cluster.single_user_name,
-                spark_env_vars=cluster.spark_env_vars,
-                spark_version=cluster.spark_version,
-                is_pinned=cluster.is_pinned,
-                opts=opts,
-            )
+            f"cluster-{cluster.name}",
+            apply_policy_default_values=cluster.apply_policy_default_values,
+            autoscale=getattr(cluster.autoscale, "pulumi_args", None),
+            autotermination_minutes=cluster.autotermination_minutes,
+            cluster_id=cluster.cluster_id,
+            cluster_name=cluster.name,
+            custom_tags=cluster.custom_tags,
+            data_security_mode=cluster.data_security_mode,
+            driver_instance_pool_id=cluster.driver_instance_pool_id,
+            driver_node_type_id=cluster.driver_node_type_id,
+            enable_elastic_disk=cluster.enable_elastic_disk,
+            enable_local_disk_encryption=cluster.enable_local_disk_encryption,
+            idempotency_token=cluster.idempotency_token,
+            init_scripts=[i.pulumi_args for i in cluster.init_scripts],
+            instance_pool_id=cluster.instance_pool_id,
+            is_pinned=cluster.is_pinned,
+            libraries=[l.pulumi_args for l in cluster.libraries],
+            node_type_id=cluster.node_type_id,
+            num_workers=cluster.num_workers,
+            policy_id=cluster.policy_id,
+            runtime_engine=cluster.runtime_engine,
+            single_user_name=cluster.single_user_name,
+            spark_conf=cluster.spark_conf,
+            spark_env_vars=cluster.spark_env_vars,
+            spark_version=cluster.spark_version,
+            ssh_public_keys=cluster.ssh_public_keys,
+            opts=opts,
+        )
 
         access_controls = []
         for permission in cluster.permissions: