+++ CHANGELOG.md
@@ -5,14 +5,18 @@
 - 
 
 **Updated**
--
+- Bronze template notebook to leverage configuration file
+- Silver template notebook to leverage configuration file
 
 **Added**
-- compute.InitScript model
 - compute.Cluster model
+- compute.InitScript model
+- compute.Notebook model
+- compute.Pipeline model
 - compute.Warehouse model
 - secrets.Secret model
 - secrets.SecretScope model
+- Pipeline configuration file management
 
 **Breaking Change**
 -